+++ CHANGELOG.md
@@ -6,8 +6,12 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
-- [Address Autofill]: removed pre-fetching for suggestions
-- [Place Autocomplete]: removed pre-fetching for suggestions
+- [Address Autofill]: removed pre-fetching for suggestions.
+- [Address Autofill]: `Suggestion.coordinate` is now an optional. `init` requires an Underlying enum parameter.
+- [Address Autofill]: Added new AddressAutofill.Suggestion.Underlying enum parameter with cases for suggestion and result inputs.
+- [Place Autocomplete]: removed pre-fetching for suggestions.
+- [Place Autocomplete]: `Suggestion.coordinate` is now an optional.
+- [Place Autocomplete]: `Result.coordinate` is now an optional.
 
 ## 1.0.0-rc.8 - 2023-10-09